+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - `civis.futures.CivisFuture` for tracking future results  
 
+### Performance Enhancements
+- ``civis.io.file_to_civis`` will perform a streaming upload to Platform if the optional ``requests-toolbelt`` package is installed.
+
 ## 1.2.0 - 2017-02-08
 ### Added
 - support for multifile csv exports