+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## 1.3.0 - 2017-03-07
 ### Added
 - `civis.futures.CivisFuture` for tracking future results